apply suggestion from code review
