+++ sky/utils/ux_utils.py
@@ -1,9 +1,13 @@
 """Utility functions for UX."""
 import contextlib
 import sys
+import traceback
+from typing import Callable
 
 import rich.console as rich_console
 
+from sky import sky_logging
+from sky.utils import common_utils
 from sky.utils import env_options
 
 console = rich_console.Console()
@@ -53,3 +57,45 @@ def enable_traceback():
     sys.tracebacklimit = 1000
     yield
     sys.tracebacklimit = original_tracelimit
+
+
+class RedirectOutputForProcess:
+    """Redirect stdout and stderr to a file.
+
+    This class enabled output redirect for multiprocessing.Process.
+    Example usage:
+
+    p = multiprocessing.Process(
+        target=RedirectOutputForProcess(func, file_name).run, args=...)
+
+    This is equal to:
+
+    p = multiprocessing.Process(target=func, args=...)
+
+    Plus redirect all stdout/stderr to file_name.
+    """
+
+    def __init__(self, func: Callable, file: str) -> None:
+        self.func = func
+        self.file = file
+
+    def run(self, *args, **kwargs):
+        with open(self.file, 'w') as f:
+            sys.stdout = f
+            sys.stderr = f
+            # reconfigure logger since the logger is initialized before
+            # with previous stdout/stderr
+            sky_logging.reload_logger()
+            logger = sky_logging.init_logger(__name__)
+            # The subprocess_util.run('sky status') inside
+            # sky.execution::_execute cannot be redirect, since we cannot
+            # directly operate on the stdout/stderr of the subprocess. This
+            # is because some code in skypilot will specify the stdout/stderr
+            # of the subprocess.
+            try:
+                self.func(*args, **kwargs)
+            except Exception as e:  # pylint: disable=broad-except
+                logger.error(f'Failed to run {self.func.__name__}. '
+                             f'Details: {common_utils.format_exception(e)}\n'
+                             f'Traceback:\n{traceback.format_exc()}')
+                raise
